Chemists-ejb/src/java/model/Basket.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package model;

import java.io.Serializable;
import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.JoinColumn;
import javax.persistence.OneToOne;
import javax.persistence.Table;
import javax.xml.bind.annotation.XmlRootElement;

/**
 *
 */
@Entity
@Table(name = "Basket", catalog = "chemists", schema = "")
@XmlRootElement
public class Basket extends BaseEntity implements Serializable {
    private static final long serialVersionUID = 1L;
    
    @OneToOne(mappedBy="basket")
//    @JoinColumn(name = "owner", referencedColumnName = "id")
    private Users owner;
    @Column (name="summa", nullable=false)
    private int summa;

    public void Basket (){
        
    }
    
    
    public Users getOwner() {
        return owner;
    }

    public void setOwner(Users owner) {
        this.owner = owner;
    }

    public int getSumma() {
        return summa;
    }

    public void setSumma(int summa) {
        this.summa = summa;
    }
    
     
    @Override
    public int hashCode() {
        int hash = 0;
        hash += (id != null ? id.hashCode() : 0);
        return hash;
    }

    @Override
    public boolean equals(Object object) {
        // TODO: Warning - this method won't work in the case the id fields are not set
        if (!(object instanceof Basket)) {
            return false;
        }
        Basket other = (Basket) object;
        if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
            return false;
        }
        return true;
    }

    @Override
    public String toString() {
        return "model.Basket[ id=" + id + " ]";
    }
    
}
